Rename discount example package to complex
