+++ examples/discount-service/src/main/java/com/acme/discount/complex/InventoryGateway.java
@@ -1,4 +1,4 @@
-package com.acme.discount.example;
+package com.acme.discount.complex;
 
 import com.acme.discount.Order;
